+++ backend/http/auth.go
@@ -24,7 +24,11 @@ type authToken struct {
 type extractor []string
 
 func (e extractor) ExtractToken(r *http.Request) (string, error) {
-	token, _ := request.HeaderExtractor{"X-Auth"}.ExtractToken(r)
+	tokenObj, err := r.Cookie("auth")
+	if err != nil {
+		return "", err
+	}
+	token := tokenObj.Value
 
 	// Checks if the token isn't empty and if it contains two dots.
 	// The former prevents incompatibility with URLs that previously
@@ -48,57 +52,25 @@ func (e extractor) ExtractToken(r *http.Request) (string, error) {
 	return "", request.ErrNoTokenInRequest
 }
 
-func withUser(fn handleFunc) handleFunc {
-	return func(w http.ResponseWriter, r *http.Request, d *data) (int, error) {
-		keyFunc := func(token *jwt.Token) (interface{}, error) {
-			return d.settings.Auth.Key, nil
-		}
-
-		var tk authToken
-		token, err := request.ParseFromRequest(r, &extractor{}, keyFunc, request.WithClaims(&tk))
-
-		if err != nil || !token.Valid {
-			return http.StatusUnauthorized, nil
-		}
-
-		expired := !tk.VerifyExpiresAt(time.Now().Add(time.Hour), true)
-		updated := tk.IssuedAt != nil && tk.IssuedAt.Unix() < d.store.Users.LastUpdate(tk.User.ID)
-
-		if expired || updated {
-			w.Header().Add("X-Renew-Token", "true")
-		}
-
-		d.user, err = d.store.Users.Get(d.server.Root, tk.User.ID)
-		if err != nil {
-			return http.StatusInternalServerError, err
-		}
-		return fn(w, r, d)
-	}
-}
-
-func withAdmin(fn handleFunc) handleFunc {
-	return withUser(func(w http.ResponseWriter, r *http.Request, d *data) (int, error) {
-		if !d.user.Perm.Admin {
-			return http.StatusForbidden, nil
-		}
-
-		return fn(w, r, d)
-	})
-}
-
-var loginHandler = func(w http.ResponseWriter, r *http.Request, d *data) (int, error) {
-	auther, err := d.store.Auth.Get(d.settings.Auth.Method)
+func loginHandler(w http.ResponseWriter, r *http.Request) {
+	// Get the authentication method from the settings
+	auther, err := store.Auth.Get(config.Auth.Method)
 	if err != nil {
-		return http.StatusInternalServerError, err
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
 	}
-
-	user, err := auther.Auth(r, d.store.Users)
+	// Authenticate the user based on the request
+	user, err := auther.Auth(r, store.Users)
 	if err == os.ErrPermission {
-		return http.StatusForbidden, nil
+		http.Error(w, http.StatusText(http.StatusForbidden), http.StatusForbidden)
+		return
 	} else if err != nil {
-		return http.StatusInternalServerError, err
-	} else {
-		return printToken(w, r, d, user)
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
+	}
+	status, err := printToken(w, r, user) // Pass the data object
+	if err != nil {
+		http.Error(w, http.StatusText(status), status)
 	}
 }
 
@@ -107,54 +79,59 @@ type signupBody struct {
 	Password string `json:"password"`
 }
 
-var signupHandler = func(w http.ResponseWriter, r *http.Request, d *data) (int, error) {
+func signupHandler(w http.ResponseWriter, r *http.Request) {
 	if !settings.Config.Auth.Signup {
-		return http.StatusMethodNotAllowed, nil
+		http.Error(w, http.StatusText(http.StatusMethodNotAllowed), http.StatusMethodNotAllowed)
+		return
 	}
 
 	if r.Body == nil {
-		return http.StatusBadRequest, nil
+		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+		return
 	}
 
 	info := &signupBody{}
 	err := json.NewDecoder(r.Body).Decode(info)
 	if err != nil {
-		return http.StatusBadRequest, err
+		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+		return
 	}
 
 	if info.Password == "" || info.Username == "" {
-		return http.StatusBadRequest, nil
+		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+		return
 	}
 
-	user := users.ApplyDefaults(users.User{})
+	user := settings.ApplyUserDefaults(users.User{})
 	user.Username = info.Username
 	user.Password = info.Password
 
-	userHome, err := d.settings.MakeUserDir(user.Username, user.Scope, d.server.Root)
+	userHome, err := config.MakeUserDir(user.Username, user.Scope, config.Server.Root)
 	if err != nil {
 		log.Printf("create user: failed to mkdir user home dir: [%s]", userHome)
-		return http.StatusInternalServerError, err
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
 	}
 	user.Scope = userHome
 	log.Printf("new user: %s, home dir: [%s].", user.Username, userHome)
-	err = d.store.Users.Save(&user)
+	err = store.Users.Save(&user)
 	if err == errors.ErrExist {
-		return http.StatusConflict, err
+		http.Error(w, http.StatusText(http.StatusConflict), http.StatusConflict)
+		return
 	} else if err != nil {
-		return http.StatusInternalServerError, err
+		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
+		return
 	}
-
-	return http.StatusOK, nil
 }
 
-var renewHandler = withUser(func(w http.ResponseWriter, r *http.Request, d *data) (int, error) {
-	return printToken(w, r, d, d.user)
-})
+func renewHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
+	return printToken(w, r, d.user)
+}
 
-func printToken(w http.ResponseWriter, _ *http.Request, d *data, user *users.User) (int, error) {
+func makeSignedToken(user *users.User) (string, error) {
 	duration, err := time.ParseDuration(settings.Config.Auth.TokenExpirationTime)
 	if err != nil {
-		duration = time.Hour * 2
+		duration = time.Hour * 2 // Default duration if parsing fails
 	}
 	claims := &authToken{
 		User: *user,
@@ -164,13 +141,15 @@ func printToken(w http.ResponseWriter, _ *http.Request, d *data, user *users.Use
 			Issuer:    "File Browser",
 		},
 	}
-
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
-	signed, err := token.SignedString(d.settings.Auth.Key)
+	return token.SignedString(config.Auth.Key)
+}
+
+func printToken(w http.ResponseWriter, _ *http.Request, user *users.User) (int, error) {
+	signed, err := makeSignedToken(user)
 	if err != nil {
 		return http.StatusInternalServerError, err
 	}
-
 	w.Header().Set("Content-Type", "text/plain")
 	if _, err := w.Write([]byte(signed)); err != nil {
 		return http.StatusInternalServerError, err