+++ frontend/src/components/files/ListingItem.vue
@@ -16,6 +16,7 @@
     :data-type="type"
     :aria-label="name"
     :aria-selected="isSelected"
+    @contextmenu="onRightClick"
     @click="quickNav ? toggleClick() : itemClick($event)"
   >
     <div @click="toggleClick" :class="{ activetitle: isMaximized && isSelected }">
@@ -134,7 +135,7 @@ export default {
         modified: this.modified,
       };
 
-      return api.getPreviewURL(file, "thumb");
+      return api.getPreviewURL(file, "small");
     },
     isThumbsEnabled() {
       return enableThumbs;
@@ -157,6 +158,21 @@ export default {
     }
   },
   methods: {
+    onRightClick(event) {
+      event.preventDefault(); // Prevent default context menu
+
+      // If no items are selected, select the right-clicked item
+      if (getters.selectedCount() === 0) {
+        mutations.addSelected(this.index);
+      }
+      mutations.showHover({
+        name: "ContextMenu",
+        props: {
+          posX: event.clientX,
+          posY: event.clientY,
+        },
+      });
+    },
     handleIntersect(entries, observer) {
       entries.forEach((entry) => {
         if (entry.isIntersecting) {
@@ -265,7 +281,6 @@ export default {
       action(overwrite, rename);
     },
     itemClick(event) {
-      console.log("should say something");
       if (this.singleClick && !state.multiple) this.open();
       else this.click(event);
     },