+++ frontend/src/views/Settings.vue
@@ -5,11 +5,7 @@
         v-for="setting in settings"
         :key="setting.id + '-main'"
         :id="setting.id + '-main'"
-        :class="{
-          active: active(setting.id + '-main'),
-          clickable: !active(setting.id + '-main'),
-        }"
-        @click="!active(setting.id + '-main') && setView(setting.id + '-main')"
+        @click="setView(setting.id + '-main')"
       >
         <!-- Dynamically render the component based on the setting -->
         <component v-if="shouldShow(setting)" :is="setting.component"></component>
@@ -79,15 +75,9 @@ export default {
   },
   methods: {
     shouldShow(setting) {
-      if (state.isMobile) {
-        const perm = setting?.perm || {};
-        // Check if all keys in setting.perm exist in state.user.perm and have truthy values
-        return Object.keys(perm).every((key) => state.user.perm[key]);
-      }
-      return this.active(setting.id + "-main");
-    },
-    active(id) {
-      return state.activeSettingsView === id;
+      const perm = setting?.perm || {};
+      // Check if all keys in setting.perm exist in state.user.perm and have truthy values
+      return Object.keys(perm).every((key) => state.user.perm[key]);
     },
     setView(view) {
       if (state.activeSettingsView === view) return;
@@ -109,11 +99,9 @@ export default {
   padding-bottom: 35vh;
   width: 100%;
 }
-.settings-views > .active > .card {
+
+.settings-views .card {
   border-style: solid;
   opacity: 1;
-}
-.settings-views .card {
-  opacity: 0.3;
 }
 </style>